+++ Shopularity/angular/public-site/src/app/containers/order-history/order-history.component.html
@@ -34,6 +34,10 @@
           </div>
         </div>
       </abp-card>
+    } @empty {
+      <div class="alert alert-info" role="alert">
+        You have no orders yet. <a href="/" class="alert-link">Go to products page</a> to place an order.
+      </div>
     }
   </div>
 </abp-page>
